refactor(stock) : 디렉터리 구조, 중복 코드, 불필요한 파일 리팩터링
