+++ src/main/java/com/github/nenidan/ne_ne_challenge/domain/shop/order/domain/OrderRepository.java
@@ -2,8 +2,8 @@ package com.github.nenidan.ne_ne_challenge.domain.shop.order.domain;
 
 import java.util.List;
 
-import com.github.nenidan.ne_ne_challenge.domain.shop.order.domain.vo.OrderId;
-import com.github.nenidan.ne_ne_challenge.domain.shop.order.domain.vo.UserId;
+import com.github.nenidan.ne_ne_challenge.domain.shop.vo.OrderId;
+import com.github.nenidan.ne_ne_challenge.domain.shop.vo.UserId;
 
 public interface OrderRepository {
     Order createOrder(Order order);
