+++ src/main/java/com/github/nenidan/ne_ne_challenge/domain/shop/order/infrastructure/entity/OrderDetailEntity.java
@@ -9,12 +9,14 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.OneToOne;
+import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 @Entity
 @Getter
+@Table(name = "order_detail")
 @NoArgsConstructor
 public class OrderDetailEntity extends BaseEntity {
