+++ src/main/java/com/github/nenidan/ne_ne_challenge/domain/shop/product/infrastructure/ProductMapper.java
@@ -1,7 +1,7 @@
 package com.github.nenidan.ne_ne_challenge.domain.shop.product.infrastructure;
 
 import com.github.nenidan.ne_ne_challenge.domain.shop.product.domain.Product;
-import com.github.nenidan.ne_ne_challenge.domain.shop.product.domain.vo.ProductId;
+import com.github.nenidan.ne_ne_challenge.domain.shop.vo.ProductId;
 import com.github.nenidan.ne_ne_challenge.domain.shop.product.infrastructure.entity.ProductEntity;
 
 public class ProductMapper {
