+++ src/main/java/com/github/nenidan/ne_ne_challenge/domain/shop/product/infrastructure/ProductRepositoryImpl.java
@@ -8,7 +8,7 @@ import com.github.nenidan.ne_ne_challenge.domain.shop.exception.ShopErrorCode;
 import com.github.nenidan.ne_ne_challenge.domain.shop.exception.ShopException;
 import com.github.nenidan.ne_ne_challenge.domain.shop.product.domain.Product;
 import com.github.nenidan.ne_ne_challenge.domain.shop.product.domain.ProductRepository;
-import com.github.nenidan.ne_ne_challenge.domain.shop.product.domain.vo.ProductId;
+import com.github.nenidan.ne_ne_challenge.domain.shop.vo.ProductId;
 import com.github.nenidan.ne_ne_challenge.domain.shop.product.infrastructure.entity.ProductEntity;
 
 @Repository
@@ -29,14 +29,14 @@ public class ProductRepositoryImpl implements ProductRepository {
 
     @Override
     public Product findById(ProductId productId) {
-        ProductEntity productEntity = productJpaRepository.findById(productId.getProductId())
+        ProductEntity productEntity = productJpaRepository.findById(productId.getValue())
             .orElseThrow(() -> new ShopException(ShopErrorCode.PRODUCT_NOT_FOUND));
         return ProductMapper.toDomain(productEntity);
     }
 
     @Override
     public Product update(ProductId productId, Product product) {
-        ProductEntity productEntity = productJpaRepository.findById(productId.getProductId())
+        ProductEntity productEntity = productJpaRepository.findById(productId.getValue())
             .orElseThrow(() -> new ShopException(ShopErrorCode.PRODUCT_NOT_FOUND));
 
         productEntity.setProductName(product.getProductName());
@@ -56,7 +56,7 @@ public class ProductRepositoryImpl implements ProductRepository {
 
     @Override
     public void delete(ProductId productId) {
-        ProductEntity productEntity = productJpaRepository.findById(productId.getProductId())
+        ProductEntity productEntity = productJpaRepository.findById(productId.getValue())
             .orElseThrow(() -> new ShopException(ShopErrorCode.PRODUCT_NOT_FOUND));
         productEntity.delete();
     }
