+++ src/main/java/com/github/nenidan/ne_ne_challenge/domain/shop/vo/ProductId.java
@@ -1,0 +1,12 @@
+package com.github.nenidan.ne_ne_challenge.domain.shop.vo;
+
+import lombok.Getter;
+
+@Getter
+public class ProductId {
+    private final Long value;
+
+    public ProductId(Long value) {
+        this.value = value;
+    }
+}
